backend/requirements.txt

streamlit
langchain
openai
python-dotenv
arxiv
pymupdf
pdf2image
pytesseract
Pillow
faiss-cpu
